Fixed signature for limit.

+++ library/src/main/java/au/id/ajlane/iostreams/EmptyIOStream.java
@@ -169,7 +169,7 @@ class EmptyIOStream implements PeekableIOStream
     }
 
     @Override
-    public IOStream limit(int size)
+    public IOStream limit(long size)
     {
         if (size < 0)
         {
